style: Move out styling codes on NavigationContainer
